---------------------------------------------------------------------- Committing in OpenScada/src Modified Files:  	Makefile Makefile.am Makefile.in main.cpp tbds.cpp tbds.h  	tconfig.cpp tconfig.h tconfigelem.cpp tconfigelem.h  	tcontroller.cpp tcontroller.h tcontrollers.cpp tcontrollers.h  	terror.cpp tgrpmodule.cpp tgrpmodule.h tmessage.cpp tmessage.h  	tmodschedul.cpp tmodschedul.h tmodule.cpp tmodule.h tparam.cpp  	tparam.h tparamcontr.cpp tparamcontr.h tparams.cpp tparams.h  	ttipcontroller.cpp ttipcontroller.h tvalue.cpp tvalue.h 	tvalueelem.cpp tvalueelem.h Added Files:  	tarhives.cpp tarhives.h tguis.cpp tguis.h tkernel.cpp  	tkernel.h tprotocols.cpp tprotocols.h tspecials.cpp  	tspecials.h ttransports.cpp ttransports.h Removed Files:  	tapplication.cpp tapplication.h tarhive.cpp tarhive.h tgui.cpp  	tgui.h tprotocol.cpp tprotocol.h tspecial.cpp tspecial.h  	ttransport.cpp ttransport.h ---------------------------------------------------------------------- - make allow of created a dinamic and static librarys of OpenScada kernel; - rename tarhive.cpp<h> -> tarhives.cpp<h> - rename tprotocol.cpp<h> -> tprotocols.cpp<h> - rename ttransport.cpp<h> -> ttransports.cpp<h> - rename tgui.cpp<h> -> tguis.cpp<h> - rename tspecial.cpp<h> -> tspecials.cpp<h> - rename tapplication.cpp<h> -> tkernel.cpp<h>
git-svn-id: svn://oscada.org/trunk/OpenScada@187 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tvalueelem.cpp
@@ -1,4 +1,4 @@
-#include "tapplication.h"
+#include "tkernel.h"
 #include "tmessage.h"
 #include "tvalue.h"
 #include "tvalueelem.h"
@@ -37,17 +37,17 @@ void TValueElem::Add(unsigned int id_val, SVAL *element )
 	if(cur_pos == st_pos) { st_pos+=1; continue; }
 	elem[id_val].vals.push_back(element->vals.substr(st_pos,cur_pos-st_pos));
 	st_pos = cur_pos+1;
-    }while(cur_pos != string::npos);
+    }while(cur_pos != (int)string::npos);
     do
     {
 	cur_pos = element->n_sel.find(";",st_pos);
 	if(cur_pos == st_pos) { st_pos+=1; continue; }
 	elem[id_val].n_sel.push_back(element->n_sel.substr(st_pos,cur_pos-st_pos));
 	st_pos = cur_pos+1;
-    }while(cur_pos != string::npos);
+    }while(cur_pos != (int)string::npos);
     elem[id_val].l_buf = element->l_buf;
 
-    for(int i_val=0;i_val < value.size();i_val++) 
+    for(unsigned i_val=0;i_val < value.size();i_val++) 
 	value[i_val]->AddElem(id_val);
 }
 
@@ -55,7 +55,7 @@ void TValueElem::Del(unsigned int id_val)
 {
     if(id_val >= Size()) throw TError("%s: id error!",o_name);
     
-    for(int i_val=0;i_val < value.size();i_val++)
+    for(unsigned i_val=0;i_val < value.size();i_val++)
 	value[i_val]->DelElem(id_val);    
     elem.erase(elem.begin()+id_val);
 }
